move purchases endpoint into user api
