+++ packages/common/src/api/index.ts
@@ -6,4 +6,3 @@ export * from './developerApps'
 export * from './suggestedTracks'
 export * from './trending'
 export * from './library'
-export * from './transactions'
